Only clean dir if not fresh

+++ generator/src/test/java/groove/behaviortransformer/BehaviorToGrooveTransformerTestHelper.java
@@ -47,7 +47,10 @@ public abstract class BehaviorToGrooveTransformerTestHelper {
         String modelName = behavior.getName();
         BehaviorToGrooveTransformer transformer = new BehaviorToGrooveTransformer();
         File outputDir = new File(this.getOutputPathIncludingSubFolder());
-        FileUtils.cleanDirectory(outputDir);
+        if (!outputDir.mkdir()) {
+            // Clean dir if not fresh
+            FileUtils.cleanDirectory(outputDir);
+        }
         transformer.generateGrooveGrammar(behavior, outputDir);
 
         // assert
